+++ wabi/src/main/kotlin/com/wap/wabi/auth/config/SecurityConfig.kt
@@ -1,6 +1,6 @@
 package com.wap.wabi.auth.config
 
-import com.wap.wabi.auth.jwt.JwtAuthenticationFilter
+import com.wap.wabi.auth.admin.service.jwt.JwtAuthenticationFilter
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.security.authentication.AuthenticationManager
